+++ enterprise/api_client/enterprise_catalog.py
@@ -26,7 +26,7 @@ class EnterpriseCatalogApiClient(JwtLmsApiClient):
     """
 
     API_BASE_URL = settings.ENTERPRISE_CATALOG_INTERNAL_ROOT_URL + '/api/v1/'
-    ENTERPRISE_CATALOG_ENDPOINT = 'enterprise-catalog'
+    ENTERPRISE_CATALOG_ENDPOINT = 'enterprise-catalogs'
     GET_CONTENT_METADATA_ENDPOINT = ENTERPRISE_CATALOG_ENDPOINT + '/{}/get_content_metadata.json'
     APPEND_SLASH = True