+++ enterprise/models.py
@@ -53,6 +53,7 @@ from enterprise.utils import (
     NotConnectedToOpenEdX,
     get_configuration_value,
     get_ecommerce_worker_user,
+    get_enterprise_worker_user,
 )
 from enterprise.validators import validate_hex_color, validate_image_extension, validate_image_size
 
@@ -786,7 +787,8 @@ class EnterpriseCustomerUser(TimeStampedModel):
         user = self.user
         identity_provider = self.enterprise_customer.identity_provider
         if user and identity_provider:
-            client = ThirdPartyAuthApiClient()
+            enterprise_worker = get_enterprise_worker_user()
+            client = ThirdPartyAuthApiClient(enterprise_worker)
             return client.get_remote_id(self.enterprise_customer.identity_provider, user.username)
         return None
 
@@ -1098,16 +1100,6 @@ class EnterpriseCustomerBrandingConfiguration(TimeStampedModel):
         verbose_name = _("Branding Configuration")
         verbose_name_plural = _("Branding Configurations")
         ordering = ['created']
-
-    def save(self, *args, **kwargs):  # pylint: disable=arguments-differ
-        """Save the enterprise customer branding config."""
-        if self.pk is None:
-            logo_image = self.logo
-            self.logo = None
-            super(EnterpriseCustomerBrandingConfiguration, self).save(*args, **kwargs)
-            self.logo = logo_image
-
-        super(EnterpriseCustomerBrandingConfiguration, self).save(*args, **kwargs)
 
     def __str__(self):
         """