+++ integrated_channels/xapi/client.py
@@ -49,7 +49,14 @@ class EnterpriseXAPIClient:
         Raises:
             ClientError: If xAPI statement fails to save.
         """
+        LOGGER.info(
+            "[Integrated Channel][xAPI] Sending statement %s to LRS endpoint %s ",
+            statement.to_json(),
+            self.lrs.endpoint
+        )
         response = self.lrs.save_statement(statement)
 
         if not response:
             raise ClientError('EnterpriseXAPIClient request failed.')
+
+        return response