+++ test_utils/__init__.py
@@ -116,7 +116,12 @@ def update_course_run_with_enterprise_context(course_run, add_utm_info=True, ent
     return course_run
 
 
-def update_course_with_enterprise_context(course, add_utm_info=True, enterprise_catalog_uuid=None):
+def update_course_with_enterprise_context(
+        course,
+        add_utm_info=True,
+        enterprise_catalog_uuid=None,
+        add_active_info=True
+):
     """
     Populate a fake course response with any necessary Enterprise context for testing purposes.
     """
@@ -135,6 +140,9 @@ def update_course_with_enterprise_context(course, add_utm_info=True, enterprise_
     )
 
     course_runs = course.get('course_runs', [])
+
+    if add_active_info:
+        course['active'] = utils.has_course_run_available_for_enrollment(course_runs)
     for course_run in course_runs:
         update_course_run_with_enterprise_context(
             course_run,
@@ -167,7 +175,8 @@ def update_program_with_enterprise_context(program, add_utm_info=True, enterpris
         update_course_with_enterprise_context(
             course,
             add_utm_info=add_utm_info,
-            enterprise_catalog_uuid=enterprise_catalog_uuid
+            enterprise_catalog_uuid=enterprise_catalog_uuid,
+            add_active_info=False,
         )
 
     return program